requirements.txt
pandas
sqlalchemy
psycopg2-binary
python-dotenv
streamlit
plotly
matplotlib
seaborn
requests
boto3 
python-dotenv
